backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
decompyle3==3.9.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.135.1
greenlet==3.3.2
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.2
python-jose==3.5.0
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
spark-parser==1.9.0
SQLAlchemy==2.0.48
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchfiles==1.1.1
websockets==16.0
xdis==6.1.7